+++ lib/src/platform/web/file_picker_web.dart
@@ -64,6 +64,7 @@ class FilePickerWeb extends FilePickerPlatform {
     bool cancelUploadOnWindowBlur = true,
     int compressionQuality = 0,
     AndroidSAFOptions? androidSafOptions,
+    bool withPersistentAccess = false,
   }) async {
     if (type != FileType.custom && (allowedExtensions?.isNotEmpty ?? false)) {
       throw Exception(
@@ -223,6 +224,8 @@ class FilePickerWeb extends FilePickerPlatform {
     List<String>? allowedExtensions,
     Uint8List? bytes,
     String? sourcePath,
+    String? sourceIdentifier,
+    String? sourcePersistentIdentifier,
     Function(FilePickerStatus)? onFileLoading,
     bool lockParentWindow = false,
   }) async {
@@ -238,7 +241,13 @@ class FilePickerWeb extends FilePickerPlatform {
       );
     }
 
-    final sourceUrl = sourcePath?.isNotEmpty == true ? sourcePath : null;
+    final sourceUrl = sourcePath?.isNotEmpty == true
+        ? sourcePath
+        : sourceIdentifier?.isNotEmpty == true
+        ? sourceIdentifier
+        : sourcePersistentIdentifier?.isNotEmpty == true
+        ? sourcePersistentIdentifier
+        : null;
 
     if (bytes == null && sourceUrl == null) {
       throw ArgumentError(