+++ lib/src/platform/windows/file_picker_windows.dart
@@ -40,6 +40,7 @@ class FilePickerWindows extends FilePickerPlatform {
     int compressionQuality = 0,
     bool cancelUploadOnWindowBlur = true,
     AndroidSAFOptions? androidSafOptions,
+    bool withPersistentAccess = false,
   }) async {
     final port = ReceivePort();
     await Isolate.spawn(
@@ -195,6 +196,8 @@ class FilePickerWindows extends FilePickerPlatform {
     List<String>? allowedExtensions,
     Uint8List? bytes,
     String? sourcePath,
+    String? sourceIdentifier,
+    String? sourcePersistentIdentifier,
     Function(FilePickerStatus)? onFileLoading,
     bool lockParentWindow = false,
   }) async {